more refactoring of importants to avoid loading errors
